display image rather than N curves for 3D, move action&toolbar to PlotTools
